+++ src/elements/three-gltf-asset.ts
@@ -1,7 +1,9 @@
 import { Group } from "three"
-import { GLTFLoader } from "three/examples/jsm/loaders/GLTFLoader"
+import { GLTF, GLTFLoader } from "three/examples/jsm/loaders/GLTFLoader"
 import { ThreeElement } from "../ThreeElement"
 import { registerElement } from "../util/registerElement"
+
+const loadedUrls: Record<string, GLTF> = {}
 
 export class ThreeGLTFAsset extends ThreeElement.for(Group) {
   static get observedAttributes() {
@@ -13,18 +15,34 @@ export class ThreeGLTFAsset extends ThreeElement.for(Group) {
   }
 
   public set url(url) {
-    const loader = new GLTFLoader()
-
     if (url) {
-      /*
-      TODO: remove previously added GLTF to handle the case where the
-      url attribute is changing
-      */
-      loader.load(url, (gltf) => {
-        this.object!.add(gltf.scene)
-        this.game.requestFrame()
-      })
+      if (url in loadedUrls) {
+        this.setupGLTF(loadedUrls[url])
+      } else {
+        const loader = new GLTFLoader()
+        loader.load(url, (gltf) => {
+          loadedUrls[url] = gltf
+          this.setupGLTF(gltf)
+        })
+      }
     }
+  }
+
+  private setupGLTF(gltf: GLTF) {
+    /* Create a copy of the GLTF just for this element */
+    const scene = gltf.scene.clone(true)
+
+    /* Apply shadow settings */
+    scene.traverse((o3d) => {
+      o3d.castShadow = this.object!.castShadow
+      o3d.receiveShadow = this.object!.receiveShadow
+    })
+
+    /* Add the GLTF to our local group */
+    this.object!.add(scene)
+
+    /* And make sure a frame will be rendered */
+    this.game.requestFrame()
   }
 }